# packages/server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: '@resala/server'
  version: 1.0.0
  description: Resala server
  license:
    name: ISC
  contact:
    name: Hossam Hamdy
paths:
...
  /api/v1/stocks/{stockId}:
    get:
      operationId: Get
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStockResponse'
      tags:
        - Stock
      security: []
      parameters:
        - in: path
          name: stockId
          required: true
          schema:
            type: string
    put:
      operationId: Update
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStockResponse'
      tags:
        - Stock
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: stockId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                quantity:
                  type: integer
                  format: int32
                sizeId:
                  type: integer
                  format: int32
                productId:
                  type: integer
                  format: int32
                colorId:
                  type: integer
                  format: int32
              type: object
    delete:
      operationId: Delete
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteStockResponse'
      tags:
        - Stock
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: stockId
          required: true
          schema:
            type: string
  /api/v1/stocks:
    get:
      operationId: List
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListStocksResponse'
      tags:
        - Stock
      security: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            default: 1
            format: int32
            type: integer
        - in: query
          name: limit
          required: false
          schema:
            default: 10
            format: int32
            type: integer
        - in: query
          name: search
          required: false
          schema:
            type: string
        - in: query
          name: productId
          required: false
          schema:
            format: int32
            type: integer
    post:
      operationId: Create
      responses:
        '201':
          description: Stock created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateStockResponse'
      tags:
        - Stock
      security:
        - jwt_auth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                quantity:
                  type: integer
                  format: int32
                sizeId:
                  type: integer
                  format: int32
                productId:
                  type: integer
                  format: int32
                colorId:
                  type: integer
                  format: int32
              type: object
  /api/v1/sizes/{sizeId}:
    get:
      operationId: Get
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSizeResponse'
      tags:
        - Size
      security: []
      parameters:
        - in: path
          name: sizeId
          required: true
          schema:
            type: string
    put:
      operationId: Update
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSizeResponse'
      tags:
        - Size
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: sizeId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              type: object
    delete:
      operationId: Delete
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSizeResponse'
      tags:
        - Size
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: sizeId
          required: true
          schema:
            type: string
  /api/v1/sizes:
    get:
      operationId: List
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSizesResponse'
      tags:
        - Size
      security: []
      parameters: []
    post:
      operationId: Create
      responses:
        '201':
          description: Size created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSizeResponse'
      tags:
        - Size
      security:
        - jwt_auth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              type: object
...
  /api/v1/images:
    get:
      operationId: List
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListImagesResponse'
      tags:
        - Image
      security:
        - jwt_auth: []
      parameters:
        - in: query
          name: productId
          required: false
          schema:
            format: int32
            type: integer
        - in: query
          name: colorId
          required: false
          schema:
            format: int32
            type: integer
    post:
      operationId: Create
      responses:
        '201':
          description: Image created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateImageResponse'
      tags:
        - Image
      security:
        - jwt_auth: []
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                colorId:
                  type: string
                productId:
                  type: string
                images:
                  items:
                    type: string
                    format: binary
                  type: array
              required:
                - colorId
                - productId
                - images
  /api/v1/images/{imageId}:
    patch:
      operationId: Update
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateImageResponse'
      tags:
        - Image
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                isPrimary:
                  type: boolean
              type: object
    delete:
      operationId: Delete
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateImageResponse'
      tags:
        - Image
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: imageId
          required: true
          schema:
            type: string
...
  /api/v1/addresses:
    get:
      operationId: ListUserAddress
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAddressResponse'
      tags:
        - Address
      security:
        - jwt_auth: []
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
    post:
      operationId: CreateUserAddress
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAddressResponse'
      tags:
        - Address
      security:
        - jwt_auth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                address:
                  type: string
                floor:
                  type: integer
                  format: int32
                building:
                  type: string
                country:
                  type: string
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                lastName:
                  type: string
                firstName:
                  type: string
                phone:
                  type: string
                userId:
                  type: string
              type: object
  /api/v1/addresses/{addressId}:
    put:
      operationId: UpdateUserAddress
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAddressResponse'
      tags:
        - Address
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                address:
                  type: string
                floor:
                  type: integer
                  format: int32
                building:
                  type: string
                country:
                  type: string
                street:
                  type: string
                city:
                  type: string
                state:
                  type: string
                lastName:
                  type: string
                firstName:
                  type: string
                phone:
                  type: string
                userId:
                  type: string
              type: object
    delete:
      operationId: DeleteUserAddress
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAddressResponse'
      tags:
        - Address
      security:
        - jwt_auth: []
      parameters:
        - in: path
          name: addressId
          required: true
          schema:
            type: string
        - in: query
          name: userId
          required: true
          schema:
            type: string
components:
...
  schemas:
    DefaultResponseBody:
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - success
      type: object
...
    Pagination:
      properties:
        total:
          type: integer
          format: int32
        limit:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
      required:
        - total
        - limit
        - page
      type: object
...
    Decimal:
      type: string
    Product:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        imageKey:
          type: string
        price:
          $ref: '#/components/schemas/Decimal'
        enDescription:
          type: string
        arDescription:
          type: string
        enName:
          type: string
        arName:
          type: string
        categoryId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
      required:
        - updatedAt
        - createdAt
        - imageUrl
        - imageKey
        - price
        - enDescription
        - arDescription
        - enName
        - arName
        - categoryId
        - id
      type: object
    Color:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        enName:
          type: string
        arName:
          type: string
        code:
          type: string
        id:
          type: integer
          format: int32
      required:
        - updatedAt
        - createdAt
        - enName
        - arName
        - code
        - id
      type: object
...
    Omit_Stock.id-or-colorId-or-productId_:
      $ref: "#/components/schemas/Pick_Stock.Exclude_keyofStock.id-or-colorId-or-prod\
        uctId__"
      description: Construct a type with the properties of T except for those in type K.
...
    Omit_Image.colorId-or-productId_:
      $ref: "#/components/schemas/Pick_Image.Exclude_keyofImage.colorId-or-productId_\
        _"
      description: Construct a type with the properties of T except for those in type K.
    GetStockResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              properties:
                images:
                  items:
                    $ref: '#/components/schemas/Omit_Image.colorId-or-productId_'
                  type: array
                sizes:
                  items:
                    allOf:
                      - $ref: '#/components/schemas/Omit_Stock.id-or-colorId-or-productId_'
                      - properties:
                          size:
                            type: string
                          stockId:
                            type: integer
                            format: int32
                        required:
                          - size
                          - stockId
                        type: object
                  type: array
                color:
                  $ref: '#/components/schemas/Color'
                product:
                  $ref: '#/components/schemas/Product'
              required:
                - images
                - sizes
                - color
                - product
              type: object
          required:
            - data
          type: object
    ListStocksResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              properties:
                stocks:
                  items:
                    properties:
                      images:
                        items:
                          $ref: '#/components/schemas/Omit_Image.colorId-or-productId_'
                        type: array
                      sizes:
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Omit_Stock.id-or-colorId-or-productId_'
                            - properties:
                                size:
                                  type: string
                                stockId:
                                  type: integer
                                  format: int32
                              required:
                                - size
                                - stockId
                              type: object
                        type: array
                      color:
                        $ref: '#/components/schemas/Color'
                      product:
                        $ref: '#/components/schemas/Product'
                    required:
                      - images
                      - sizes
                      - color
                      - product
                    type: object
                  type: array
                pagination:
                  $ref: '#/components/schemas/Pagination'
              required:
                - stocks
                - pagination
              type: object
          required:
            - data
          type: object
    Stock:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        quantity:
          type: integer
          format: int32
        sizeId:
          type: integer
          format: int32
        colorId:
          type: integer
          format: int32
        productId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
      required:
        - updatedAt
        - createdAt
        - quantity
        - sizeId
        - colorId
        - productId
        - id
      type: object
    CreateStockResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/Stock'
          required:
            - data
          type: object
    UpdateStockResponse:
      $ref: '#/components/schemas/CreateStockResponse'
    DeleteStockResponse:
      $ref: '#/components/schemas/CreateStockResponse'
    Size:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        id:
          type: integer
          format: int32
      required:
        - updatedAt
        - createdAt
        - name
        - id
      type: object
    GetSizeResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/Size'
          required:
            - data
          type: object
    ListSizesResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              items:
                $ref: '#/components/schemas/Size'
              type: array
          required:
            - data
          type: object
    CreateSizeResponse:
      $ref: '#/components/schemas/GetSizeResponse'
    UpdateSizeResponse:
      $ref: '#/components/schemas/GetSizeResponse'
    DeleteSizeResponse:
      $ref: '#/components/schemas/GetSizeResponse'
...
    Image:
      properties:
        createdAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        imageKey:
          type: string
        isPrimary:
          type: boolean
        productId:
          type: integer
          format: int32
        colorId:
          type: integer
          format: int32
        id:
          type: integer
          format: int32
      required:
        - createdAt
        - imageUrl
        - imageKey
        - isPrimary
        - productId
        - colorId
        - id
      type: object
...
    Address:
      properties:
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        lastName:
          type: string
        firstName:
          type: string
        phone:
          type: string
        address:
          type: string
          nullable: true
        floor:
          type: integer
          format: int32
          nullable: true
        building:
          type: string
          nullable: true
        street:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        id:
          type: integer
          format: int32
      required:
        - updatedAt
        - createdAt
        - lastName
        - firstName
        - phone
        - address
        - floor
        - building
        - street
        - city
        - state
        - country
        - id
      type: object
...
    ListImagesResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              items:
                $ref: '#/components/schemas/Image'
              type: array
          required:
            - data
          type: object
    CreateImageResponse:
      $ref: '#/components/schemas/DefaultResponseBody'
    UpdateImageResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/Image'
          required:
            - data
          type: object
...
    ListAddressResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              items:
                $ref: '#/components/schemas/Address'
              type: array
          required:
            - data
          type: object
    GetAddressResponse:
      allOf:
        - $ref: '#/components/schemas/DefaultResponseBody'
        - properties:
            data:
              $ref: '#/components/schemas/Address'
          required:
            - data
          type: object
    CreateAddressResponse:
      $ref: '#/components/schemas/GetAddressResponse'
    UpdateAddressResponse:
      $ref: '#/components/schemas/GetAddressResponse'
    DeleteAddressResponse:
      $ref: '#/components/schemas/GetAddressResponse'
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.resala.live
    description: Development server